requirements.txt
Flask
Pillow
numpy
gunicorn

# IR AGREGANDO
